requirements.txt
requests
beautifulsoup4
sqlalchemy
openai
python-dotenv
pyyaml
loguru
